# [libc++] Optimize std::find for segmented iterators