@EricWF 2023-10-20
The benchmark results are great. Could you also provide a description of how this optimization works?